+++ tests/e2e/bcd_consumer_integration_test.go
@@ -477,7 +477,7 @@ func (s *BCDConsumerIntegrationTestSuite) Test08ActivateDelegation() {
 	s.Equal(hex.EncodeToString(activeDel.StakingTx), hex.EncodeToString(dataFromContract.Delegations[0].StakingTx))
 	s.Equal(activeDel.SlashingTx.ToHexStr(), hex.EncodeToString(dataFromContract.Delegations[0].SlashingTx))
 
-	// Query and assert finality provider voting power is equal to the total stake
+	// Query and assert finality provider has total active stake
 	s.Eventually(func() bool {
 		fpInfo, err := s.cosmwasmController.QueryFinalityProviderInfo(consumerFp.FinalityProvider.BtcPk.MustToBTCPK())
 		if err != nil {
@@ -485,7 +485,18 @@ func (s *BCDConsumerIntegrationTestSuite) Test08ActivateDelegation() {
 			return false
 		}
 
-		return fpInfo != nil && fpInfo.Power == activeDel.TotalSat && fpInfo.BtcPkHex == consumerFp.FinalityProvider.BtcPk.MarshalHex()
+		return fpInfo != nil &&
+			fpInfo.TotalActiveSats == activeDel.TotalSat &&
+			fpInfo.BtcPkHex == consumerFp.FinalityProvider.BtcPk.MarshalHex() &&
+			!fpInfo.Slashed
+	}, time.Minute, time.Second*5)
+	// Query and assert finality provider has voting power
+	s.Eventually(func() bool {
+		height, err := s.cosmwasmController.QueryLatestBlockHeight()
+		s.NoError(err)
+		fpHasPower, err := s.cosmwasmController.QueryFinalityProviderHasPower(consumerFp.FinalityProvider.BtcPk.MustToBTCPK(), height)
+		s.NoError(err)
+		return fpHasPower
 	}, time.Minute, time.Second*5)
 }
 
@@ -661,15 +672,13 @@ func (s *BCDConsumerIntegrationTestSuite) Test10BabylonFPCascadedSlashing() {
 	s.Require().NoError(err)
 	s.Require().NotNil(consumerFp)
 
-	// query and assert finality provider voting power is zero after slashing
+	// query and assert finality provider has zero voting power
 	s.Eventually(func() bool {
-		fpInfo, err := s.cosmwasmController.QueryFinalityProviderInfo(consumerFp.FinalityProvider.BtcPk.MustToBTCPK())
-		if err != nil {
-			s.T().Logf("Error querying finality providers by power: %v", err)
-			return false
-		}
-
-		return fpInfo != nil && fpInfo.Power == 0 && fpInfo.BtcPkHex == consumerFp.FinalityProvider.BtcPk.MarshalHex()
+		height, err := s.cosmwasmController.QueryLatestBlockHeight()
+		s.NoError(err)
+		fpHasPower, err := s.cosmwasmController.QueryFinalityProviderHasPower(consumerFp.FinalityProvider.BtcPk.MustToBTCPK(), height)
+		s.NoError(err)
+		return !fpHasPower
 	}, time.Minute, time.Second*5)
 }
 
@@ -701,25 +710,10 @@ func (s *BCDConsumerIntegrationTestSuite) Test11ConsumerFPCascadedSlashing() {
 		return err == nil && dataFromContract != nil && len(dataFromContract.Delegations) == 2
 	}, time.Second*30, time.Second)
 
-	// query and assert consumer finality provider's voting power is equal to the total stake
-	s.Eventually(func() bool {
-		fpInfo, err := s.cosmwasmController.QueryFinalityProviderInfo(consumerFp.FinalityProvider.BtcPk.MustToBTCPK())
-		if err != nil {
-			s.T().Logf("Error querying finality provider info: %v", err)
-			return false
-		}
-
-		return fpInfo != nil && fpInfo.Power == delegation.TotalSat && fpInfo.BtcPkHex == consumerFp.FinalityProvider.BtcPk.MarshalHex()
-	}, time.Minute, time.Second*5)
-
 	// get the latest block height and block on the consumer chain
 	consumerNodeStatus, err := s.cosmwasmController.GetCometNodeStatus()
 	s.NoError(err)
-	s.NotNil(consumerNodeStatus)
 	consumerLatestBlockHeight := uint64(consumerNodeStatus.SyncInfo.LatestBlockHeight)
-	consumerLatestBlock, err := s.cosmwasmController.QueryIndexedBlock(consumerLatestBlockHeight)
-	s.NoError(err)
-	s.NotNil(consumerLatestBlock)
 
 	// commit public randomness at the latest block height on the consumer chain
 	randListInfo, msgCommitPubRandList, err := datagen.GenRandomMsgCommitPubRandList(r, consumerFpBTCSK2, s.cosmwasmController.GetFpRandCommitContext(), consumerLatestBlockHeight, 200)
@@ -732,21 +726,36 @@ func (s *BCDConsumerIntegrationTestSuite) Test11ConsumerFPCascadedSlashing() {
 	// finalize the consumer chain until the latest block height
 	s.finalizeUntilConsumerHeight(consumerLatestBlockHeight)
 
+	// query and assert consumer finality provider has voting power
+	s.Eventually(func() bool {
+		height, err := s.cosmwasmController.QueryLatestBlockHeight()
+		s.NoError(err)
+		fpHasPower, err := s.cosmwasmController.QueryFinalityProviderHasPower(consumerFp.FinalityProvider.BtcPk.MustToBTCPK(), height)
+		s.NoError(err)
+		return fpHasPower
+	}, time.Minute, time.Second*5)
+
+	// get the latest block height and block on the consumer chain
+	consumerNodeStatus, err = s.cosmwasmController.GetCometNodeStatus()
+	s.NoError(err)
+	consumerLatestBlockHeight2 := uint64(consumerNodeStatus.SyncInfo.LatestBlockHeight)
+
 	// Consumer finality provider submits finality signature
+	idx := consumerLatestBlockHeight2 - consumerLatestBlockHeight
 	_, err = s.cosmwasmController.SubmitFinalitySig(
 		consumerFpBTCSK2,
 		consumerFpBTCPK2,
-		randListInfo.SRList[0],
-		&randListInfo.PRList[0],
-		randListInfo.ProofList[0].ToProto(),
-		consumerLatestBlockHeight,
+		randListInfo.SRList[idx],
+		&randListInfo.PRList[idx],
+		randListInfo.ProofList[idx].ToProto(),
+		consumerLatestBlockHeight2,
 	)
 	s.NoError(err)
 	s.T().Logf("Finality sig for height %d was submitted successfully", consumerLatestBlockHeight)
 
 	// ensure consumer finality provider's finality signature is received and stored in the smart contract
 	s.Eventually(func() bool {
-		fpSigsResponse, err := s.cosmwasmController.QueryFinalitySignature(consumerFp.FinalityProvider.BtcPk.MarshalHex(), uint64(consumerLatestBlockHeight))
+		fpSigsResponse, err := s.cosmwasmController.QueryFinalitySignature(consumerFp.FinalityProvider.BtcPk.MarshalHex(), consumerLatestBlockHeight2)
 		if err != nil {
 			s.T().Logf("failed to query finality signature: %s", err.Error())
 			return false
@@ -762,15 +771,15 @@ func (s *BCDConsumerIntegrationTestSuite) Test11ConsumerFPCascadedSlashing() {
 		r,
 		consumerFpBTCSK2,
 		consumerFpBTCPK2,
-		randListInfo.SRList[0],
-		&randListInfo.PRList[0],
-		randListInfo.ProofList[0].ToProto(),
-		consumerLatestBlockHeight,
+		randListInfo.SRList[idx],
+		&randListInfo.PRList[idx],
+		randListInfo.ProofList[idx].ToProto(),
+		consumerLatestBlockHeight2,
 	)
 	s.NoError(err)
 	s.NotNil(txResp)
 
-	// ensure consumer finality provider is slashed and has zero voting power
+	// ensure consumer finality provider is slashed
 	s.Eventually(func() bool {
 		fp, err := s.cosmwasmController.QueryFinalityProvider(consumerFp.FinalityProvider.BtcPk.MarshalHex())
 		if err != nil || fp == nil || fp.SlashedHeight == 0 {
@@ -782,7 +791,17 @@ func (s *BCDConsumerIntegrationTestSuite) Test11ConsumerFPCascadedSlashing() {
 			return false
 		}
 
-		return fpInfo != nil && fpInfo.Power == 0 && fpInfo.BtcPkHex == consumerFp.FinalityProvider.BtcPk.MarshalHex()
+		return fpInfo != nil &&
+			fpInfo.BtcPkHex == consumerFp.FinalityProvider.BtcPk.MarshalHex() &&
+			fpInfo.Slashed
+	}, time.Minute, time.Second*5)
+	// query and assert consumer finality provider has zero voting power
+	s.Eventually(func() bool {
+		height, err := s.cosmwasmController.QueryLatestBlockHeight()
+		s.NoError(err)
+		fpHasPower, err := s.cosmwasmController.QueryFinalityProviderHasPower(consumerFp.FinalityProvider.BtcPk.MustToBTCPK(), height)
+		s.NoError(err)
+		return !fpHasPower
 	}, time.Minute, time.Second*5)
 
 	// check the babylon finality provider's voting power is discounted (cascaded slashing)
@@ -858,14 +877,17 @@ func (s *BCDConsumerIntegrationTestSuite) Test12ConsumerDelegationExpiry() {
 		return err == nil && dataFromContract != nil
 	}, time.Second*30, time.Second)
 
-	// query and assert consumer finality provider's voting power is equal to the total stake
+	// query and assert consumer finality provider's total stake
 	s.Eventually(func() bool {
 		fpInfo, err := s.cosmwasmController.QueryFinalityProviderInfo(consumerFp.FinalityProvider.BtcPk.MustToBTCPK())
 		if err != nil {
 			s.T().Logf("Error querying finality provider info: %v", err)
 			return false
 		}
-		return fpInfo != nil && fpInfo.Power == delegation.TotalSat && fpInfo.BtcPkHex == consumerFp.FinalityProvider.BtcPk.MarshalHex()
+		return fpInfo != nil &&
+			fpInfo.TotalActiveSats == delegation.TotalSat &&
+			fpInfo.BtcPkHex == consumerFp.FinalityProvider.BtcPk.MarshalHex() &&
+			!fpInfo.Slashed
 	}, time.Minute, time.Second*5)
 
 	// just enough headers to exceed the delegation's end height
@@ -874,10 +896,13 @@ func (s *BCDConsumerIntegrationTestSuite) Test12ConsumerDelegationExpiry() {
 		s.Require().NoError(err)
 	}
 
-	// query and assert consumer finality provider's voting power is zero after inserting headers
+	// query and assert consumer finality provider's total stake is zero after inserting headers
 	s.Eventually(func() bool {
 		fpInfo, err := s.cosmwasmController.QueryFinalityProviderInfo(consumerFp.FinalityProvider.BtcPk.MustToBTCPK())
-		return err == nil && fpInfo != nil && fpInfo.Power == 0 && fpInfo.BtcPkHex == consumerFp.FinalityProvider.BtcPk.MarshalHex()
+		return err == nil && fpInfo != nil &&
+			fpInfo.TotalActiveSats == 0 &&
+			fpInfo.BtcPkHex == consumerFp.FinalityProvider.BtcPk.MarshalHex() &&
+			!fpInfo.Slashed
 	}, time.Minute, time.Second*5)
 }